monami/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
django-environ==0.4.4
django-filter==1.1.0
django-webpack-loader==0.5.0
djangorestframework==3.7.7
httpie==0.9.9
idna==2.6
Markdown==2.6.10
psycopg2==2.7.3.2
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
